Raise ResponseError unless response is valid
When openlogi returns a request_invalid error it means that the request
was invalid, e.g. because the access token was invalid, and in this case
we want to raise an error rather than return anything.

For other errors (e.g. data validation errors) we want to return an
actual object with errors on it.